DAG problem.
Had a better understanding of DP. DP has to implementations, top-down
and bottom-up.

To use bottom-up, either we know a topological sort by heart, for
example, in coin change problem the topo sort is from 1 to n, or we do
explicit topo sort and use this sort to update information, for
example, in DAG.

However, if the topo sort is hard to implement, or not practical due to
complicated states relations, etc. Then we can use top-down DP. The
top-down DP does topological sort implicitly.

One lesson learned today: global variable is certainly not a good idea
in industry. Even in competitive programming area, it may lead to
problems. Using vector, stack, queue, etc. are fine. They are large
structure variables which are not easy to be forgotten to initialise.
But when using global int variables, the chance of forgetting
initialise is much higher!! Try to avoid using global int variables.